LogQueueFifo: take a reference on logqueue, while the finish callback was not executed
Otherwise it could cause a use-after-free,
when the file destination closes the logwriter at time reap (possible race condition)

There were some core files at user,
and in these core files shows the symptoms of a use-after-free incident

This correction is based on only a guess.
Anyway in the logqueue-fifo, there was nothing which guarantee,
that the queue won't be freed before the callback is called.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>