LogSource: in some cases we have to know, that the last message is sent
It happens when the current window size will be equal with the
initial window size. That means no more message is on the logpath.

In that case the LogReader sends a NC_LAST_MSG_SENT notification
to its control pipe.

E.g.: The file reader can be informed, that the last message of the file
is fully processed.

Signed-off-by: Juhász Viktor <juhasz.viktor81@gmail.com>