Wildcard: Handle the file deletion event correctly.
If an actually read file was deleted, the wildcard file source didn't release
that file, reserving a slot unnecessary.

With this patch the deleted file become monitored,
and if its EOF is reached, then the polling of file will be stopped.
Finally if all the messages was sent from the monitored file,
it will be released, giving a free slot for a new file.

This monitoring is implemented by the FileStateHandler.
It processes the notifications and calls the related callback.

The file state has to be followed from beginning,
because if a file is deleted after all its messages are sent,
then the file has to be released immediately.

Signed-off-by: Juhász Viktor <juhasz.viktor81@gmail.com>