file-list: add a waiting list for file
If a file has no space in wildcard filesource (file-limit reached),
then this file is put into this list. If an active file is deleted,
and all messages are sent from that file, then a file will be pop
from this list and continue with processing of the new file.

Signed-off-by: Juhász Viktor <juhasz.viktor81@gmail.com>