directory-monitor: ensure, that delete notification comes about every file,
if a directory is deleted. Previously the directoy monitor was stopped and freed,
before it could send the notifications about file deletion.

Now when a directory monitor detects, that its monitored directory
is deleted, than it send a DIRECTORY_DELETED event type.

Avoiding use-after free situation, the directory won't be freed in
the callback function, but a destruction task will be registered.